Avancement au 30/09/2020 16:50

+++ variables12.py
@@ -3,7 +3,7 @@ items_unit_prices = input("indiquer le prix unitaire HT de l'article: ")
 items_quantity = input("indiquer le nombre d'articles dans votre panier: ")
 
 #Calcul et affichage du TOTAL HT, TVA et TOTAL TTC
-items_total_ht = int(items_unit_prices) * int(items_quantity)
+items_total_ht = float(items_unit_prices) * int(items_quantity)
 print("TOTAL HT: " + str(items_total_ht) + " EUR")
 
 items_TVA_value = items_total_ht / 100 * 20
